pyproject.toml
[project]
name = "agentic-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "loguru>=0.7.3",
    "mistralai>=1.9.7",
    "openai>=1.100.2",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
    "tenacity>=9.1.2",
]
